Simple_Document _Scanner/requirements.txt
streamlit
opencv-python-headless
numpy
easyocr
pandas
Pillow